GPL V3 -> MIT
I still am the sole owner of the work i have created apart from the GUT addon which is why i feel like it's justified to choose a license that makes sense.

Most Godot code is MIT and for a good reason, GPL is anti-proprietary and it makes no sense for a game. I love GPL, but this project benefits from MIT.