!!!REMOVED SOME METHODS!!!
Important: Methods that were just bloat or not all that useful were simply removed.
This is necessary to remember because at some point i may want to re-introduce these methods as i develop.

Hopefully the cleaning up of bloat causes an improvement in the library's quality and how it's used.